[fix] handle 404 response on api /auth/set-cookie
