+++ app/api/route/auth.ts
@@ -3,10 +3,6 @@ import { deleteCookie, setCookie } from "hono/cookie"
 
 const auth = new Hono()
 
-auth.get("/", (c) => {
-    return c.text("cok")
-})
-
 auth.get("/logout", async (c) => {
     deleteCookie(c, "firebase_token", { path: "/" })
     return c.json({
@@ -14,22 +10,25 @@ auth.get("/logout", async (c) => {
     }, 200)
 })
 
-auth.post("/", async (c) => {
-    const { token } = await c.req.json()
-
+auth.post("/set-cookie", async (c) => {
+    const { token } = await c.req.json();
     if (!token) {
         return c.json({
             message: "Token is required!"
-        }, 400)
+        }, 400);
     }
-    if (token) {
+    try {
         setCookie(c, "firebase_token", token, {
             httpOnly: true,
             secure: process.env.NODE_ENV === "production",
             sameSite: "strict",
             path: "/",
-        })
+        });
+        return c.json({ message: "Authenticated" });
+    } catch (e) {
+        console.error(e);
+        return c.json({ message: "Internal Server Error" }, 500);
     }
-})
+});
 
 export default auth
